Basic name implementation
