+++ profile/src/main/java/com/scottlogic/deg/profile/reader/CatalogService.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.profile.reader.names;
+package com.scottlogic.deg.profile.reader;
 
 import java.util.Set;
