+++ profile/src/main/java/com/scottlogic/deg/profile/reader/MainConstraintReader.java
@@ -1,5 +1,6 @@
 package com.scottlogic.deg.profile.reader;
 
+import com.google.inject.Inject;
 import com.scottlogic.deg.common.profile.constraints.Constraint;
 import com.scottlogic.deg.common.profile.RuleInformation;
 import com.scottlogic.deg.common.profile.ProfileFields;
@@ -13,8 +14,9 @@ import java.util.Set;
 public class MainConstraintReader implements ConstraintReader {
     private final AtomicConstraintReaderLookup atomicConstraintReaderLookup;
 
-    public MainConstraintReader() {
-        this.atomicConstraintReaderLookup = new AtomicConstraintReaderLookup();
+    @Inject
+    public MainConstraintReader(AtomicConstraintReaderLookup atomicConstraintReaderLookup) {
+        this.atomicConstraintReaderLookup = atomicConstraintReaderLookup;
     }
 
     @Override
