+++ profile/src/main/java/com/scottlogic/deg/profile/reader/names/NamePopulator.java
@@ -1,0 +1,8 @@
+package com.scottlogic.deg.profile.reader.names;
+
+import java.util.Set;
+
+public interface NamePopulator<T> {
+
+    Set<NameFrequencyHolder> retrieveNames(T config);
+}
